feat: enable audit in most tke components
