+++ cmd/tke-auth-api/app/options/options.go
@@ -19,6 +19,7 @@
 package options
 
 import (
+	"fmt"
 	"github.com/spf13/pflag"
 	genericapiserveroptions "k8s.io/apiserver/pkg/server/options"
 	apiserveroptions "tkestack.io/tke/pkg/apiserver/options"
@@ -37,6 +38,7 @@ type Options struct {
 	Authorization  *AuthorizationOptions
 	ETCD           *storageoptions.ETCDStorageOptions
 	Auth           *AuthOptions
+	Audit          *genericapiserveroptions.AuditOptions
 }
 
 // NewOptions creates a new Options with a default config.
@@ -50,6 +52,7 @@ func NewOptions(serverName string) *Options {
 		Authorization:  NewAuthorizationOptions(),
 		ETCD:           storageoptions.NewETCDStorageOptions("/tke/auth-api"),
 		Auth:           NewAuthOptions(),
+		Audit:          genericapiserveroptions.NewAuditOptions(),
 	}
 }
 
@@ -63,6 +66,7 @@ func (o *Options) AddFlags(fs *pflag.FlagSet) {
 	o.Authentication.AddFlags(fs)
 	o.Authorization.AddFlags(fs)
 	o.Auth.AddFlags(fs)
+	o.Audit.AddFlags(fs)
 }
 
 // ApplyFlags parsing parameters from the command line or configuration file
@@ -103,6 +107,9 @@ func (o *Options) Complete() error {
 			return err
 		}
 		o.ETCD.WatchCacheSizes = watchCacheSizes
+	}
+	if (o.Audit.WebhookOptions.ConfigFile != "" || o.Audit.LogOptions.Path != "") && o.Audit.PolicyFile == "" {
+		return fmt.Errorf("audit log/webhook config specified, but audit policy file is empty")
 	}
 
 	return nil
