+++ cmd/tke-business-api/app/options/options.go
@@ -19,6 +19,7 @@
 package options
 
 import (
+	"fmt"
 	"github.com/spf13/pflag"
 	genericapiserveroptions "k8s.io/apiserver/pkg/server/options"
 	apiserveroptions "tkestack.io/tke/pkg/apiserver/options"
@@ -40,6 +41,7 @@ type Options struct {
 	PlatformAPIClient *controlleroptions.APIServerClientOptions
 	RegistryAPIClient *controlleroptions.APIServerClientOptions
 	FeatureOptions    *FeatureOptions
+	Audit             *genericapiserveroptions.AuditOptions
 }
 
 // NewOptions creates a new Options with a default config.
@@ -55,6 +57,7 @@ func NewOptions(serverName string) *Options {
 		PlatformAPIClient: controlleroptions.NewAPIServerClientOptions("platform", true),
 		RegistryAPIClient: controlleroptions.NewAPIServerClientOptions("registry", false),
 		FeatureOptions:    NewFeatureOptions(),
+		Audit:             genericapiserveroptions.NewAuditOptions(),
 	}
 }
 
@@ -70,6 +73,7 @@ func (o *Options) AddFlags(fs *pflag.FlagSet) {
 	o.PlatformAPIClient.AddFlags(fs)
 	o.RegistryAPIClient.AddFlags(fs)
 	o.FeatureOptions.AddFlags(fs)
+	o.Audit.AddFlags(fs)
 }
 
 // ApplyFlags parsing parameters from the command line or configuration file
@@ -112,6 +116,9 @@ func (o *Options) Complete() error {
 			return err
 		}
 		o.ETCD.WatchCacheSizes = watchCacheSizes
+	}
+	if (o.Audit.WebhookOptions.ConfigFile != "" || o.Audit.LogOptions.Path != "") && o.Audit.PolicyFile == "" {
+		return fmt.Errorf("audit log/webhook config specified, but audit policy file is empty")
 	}
 	return nil
 }
