+++ cmd/tke-notify-api/app/options/options.go
@@ -19,6 +19,7 @@
 package options
 
 import (
+	"fmt"
 	"github.com/spf13/pflag"
 	genericapiserveroptions "k8s.io/apiserver/pkg/server/options"
 	apiserveroptions "tkestack.io/tke/pkg/apiserver/options"
@@ -39,6 +40,7 @@ type Options struct {
 	Authorization     *apiserveroptions.AuthorizationOptions
 	PlatformAPIClient *controlleroptions.APIServerClientOptions
 	FeatureOptions    *FeatureOptions
+	Audit             *genericapiserveroptions.AuditOptions
 }
 
 // NewOptions creates a new Options with a default config.
@@ -53,6 +55,7 @@ func NewOptions(serverName string) *Options {
 		Authorization:     apiserveroptions.NewAuthorizationOptions(),
 		PlatformAPIClient: controlleroptions.NewAPIServerClientOptions("platform", true),
 		FeatureOptions:    NewFeatureOptions(),
+		Audit:             genericapiserveroptions.NewAuditOptions(),
 	}
 }
 
@@ -67,6 +70,7 @@ func (o *Options) AddFlags(fs *pflag.FlagSet) {
 	o.Authorization.AddFlags(fs)
 	o.PlatformAPIClient.AddFlags(fs)
 	o.FeatureOptions.AddFlags(fs)
+	o.Audit.AddFlags(fs)
 }
 
 // ApplyFlags parsing parameters from the command line or configuration file
@@ -108,6 +112,9 @@ func (o *Options) Complete() error {
 			return err
 		}
 		o.ETCD.WatchCacheSizes = watchCacheSizes
+	}
+	if (o.Audit.WebhookOptions.ConfigFile != "" || o.Audit.LogOptions.Path != "") && o.Audit.PolicyFile == "" {
+		return fmt.Errorf("audit log/webhook config specified, but audit policy file is empty")
 	}
 	return nil
 }
