+++ cmd/tke-registry-api/app/config/config.go
@@ -38,6 +38,7 @@ import (
 	"tkestack.io/tke/pkg/apiserver/handler"
 	"tkestack.io/tke/pkg/apiserver/openapi"
 	"tkestack.io/tke/pkg/apiserver/storage"
+	"tkestack.io/tke/pkg/apiserver/util"
 	registryconfig "tkestack.io/tke/pkg/registry/apis/config"
 	registryconfigvalidation "tkestack.io/tke/pkg/registry/apis/config/validation"
 	"tkestack.io/tke/pkg/registry/apiserver"
@@ -104,6 +105,9 @@ func CreateConfigFromOptions(serverName string, opts *options.Options) (*Config,
 	genericAPIServerConfig.EnableIndex = false
 	genericAPIServerConfig.EnableProfiling = false
 
+	if err := util.SetupAuditConfig(genericAPIServerConfig, opts.Audit); err != nil {
+		return nil, err
+	}
 	if err := opts.Generic.ApplyTo(genericAPIServerConfig); err != nil {
 		return nil, err
 	}
