+++ cmd/tke-registry-api/app/options/options.go
@@ -19,6 +19,7 @@
 package options
 
 import (
+	"fmt"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 	genericapiserveroptions "k8s.io/apiserver/pkg/server/options"
@@ -45,6 +46,7 @@ type Options struct {
 	// The Registry will load its initial configuration from this file.
 	// The path may be absolute or relative; relative paths are under the Registry's current working directory.
 	RegistryConfig string
+	Audit          *genericapiserveroptions.AuditOptions
 }
 
 // NewOptions creates a new Options with a default config.
@@ -57,6 +59,7 @@ func NewOptions(serverName string) *Options {
 		Generic:        apiserveroptions.NewGenericOptions(),
 		Authentication: apiserveroptions.NewAuthenticationWithAPIOptions(),
 		Authorization:  apiserveroptions.NewAuthorizationOptions(),
+		Audit:          genericapiserveroptions.NewAuditOptions(),
 	}
 }
 
@@ -69,6 +72,7 @@ func (o *Options) AddFlags(fs *pflag.FlagSet) {
 	o.Generic.AddFlags(fs)
 	o.Authentication.AddFlags(fs)
 	o.Authorization.AddFlags(fs)
+	o.Audit.AddFlags(fs)
 
 	fs.String(flagRegistryConfig, o.RegistryConfig,
 		"The Registry will load its initial configuration from this file. The path may be absolute or relative; relative paths start at the Registry's current working directory. Omit this flag to use the built-in default configuration values.")
@@ -114,6 +118,9 @@ func (o *Options) Complete() error {
 			return err
 		}
 		o.ETCD.WatchCacheSizes = watchCacheSizes
+	}
+	if (o.Audit.WebhookOptions.ConfigFile != "" || o.Audit.LogOptions.Path != "") && o.Audit.PolicyFile == "" {
+		return fmt.Errorf("audit log/webhook config specified, but audit policy file is empty")
 	}
 	return nil
 }
